+++ frontend/src/components/common/verticalWishList/verticalWishList.module.css
@@ -1,0 +1,47 @@
+.VerticalWishList {
+  height: 100vh;
+}
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-bottom: 0.3px solid rgba(0, 0, 0, 0.3);
+}
+
+.closeButtonBox {
+  left: 5%;
+}
+.body {
+  height: 90%;
+  overflow: auto;
+}
+.list {
+  padding: 16px;
+  height: 100%;
+}
+
+.wishItem {
+  display: flex;
+  align-items: center;
+}
+
+.imageContainer {
+  position: relative;
+  width: 30vw;
+}
+.imageContainer::before {
+  content: "";
+  display: block;
+  padding-top: 100%; /* 정사각형일 경우 */
+}
+.image {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 12px;
+}
